feat: add types and convert to tsx
